+++ test/dummy/app/controllers/local_time_demo_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class LocalTimeDemoController < ApplicationController
-  def index
-    @now = Time.current
-  end
-end